+++ packages/terra-application-header-layout/src/ApplicationHeaderLayout.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 
-import styles from './ApplicationHeaderLayout.scss';
+import styles from './ApplicationHeaderLayout.module.scss';
 
 const cx = classNames.bind(styles);